+++ modules_k/mi_xmlrpc/abyss_xmlrpc_server.c
@@ -370,7 +370,7 @@ processCall(TSession *        const abyssSessionP,
     if (contentSize > xmlrpc_limit_get(XMLRPC_XML_SIZE_LIMIT_ID))
         xmlrpc_env_set_fault_formatted(
             &env, XMLRPC_LIMIT_EXCEEDED_ERROR,
-            "XML-RPC request too large (%d bytes)", contentSize);
+            "XML-RPC request too large (%ld bytes)", (long)contentSize);
     else {
         xmlrpc_mem_block *body=0;
         /* Read XML data off the wire. */
